MB-8445: Fix CPU utilization and other counters
The formatted WMI counters automatically calculate rates where
specified. However, ns_server is calculating rate again, and
so we need to send raw values so as to appear like sigar.

It is actually preferable to let Windows take care of sampling
by moving sampling logic from ns_server into port_sigar and
then disabling it for Windows.  But this is for a future
release. Fixing just the immediate issue for now.

Change-Id: I6ff0816ac4d2c233a0c22502d33116a3fedf3616